src/functions/translation_processor/requirements.txt
google-cloud-storage==2.14.0
google-cloud-translate==3.13.0
google-cloud-aiplatform==1.38.1
zep-cloud==1.0.2
functions-framework==3.5.0
structlog==24.1.0 